quiet warnings

+++ GG/GG/Enum.h
@@ -243,7 +243,7 @@ private:
         bool is_hex = base == 16;
 
         int retval = 0;
-        for (auto c : txt.substr((is_negative ? 1u : 0u) + 2u*is_hex)) {
+        for (auto c : txt.substr(static_cast<size_t>((is_negative ? 1u : 0u) + 2u*is_hex))) {
             retval *= base;
             std::size_t digit = valid_chars.find(c);
             retval += static_cast<int>(digit);
